Skip playing Invalid cards in the card queue.
NOTE: not sure if this is technically correct or if invalid cards are
meant to be played/processed. There was a bug when running the
SimpleAgent in the Hexaghost scenario where an Invalid card would be
discarded and trip an assertion. Not sure what the cause was.

Possibilities: broke something when isolating the battle components.
Somehow initializing the GameContext/BattleContext from the scenario
incorrectly. Found a bug in the original implementation.